requirements.txt
matplotlib==3.0.2
numpy==1.16.1
scipy==1.2.1
Pillow==6.0.0
